Refactored backend to support 3 user roles

+++ labshare-client/src/backend/endpoints/Profile.ts
@@ -1,16 +1,15 @@
 import express from "express";
-import mongoose, { Model } from 'mongoose';
 import { Validator as v } from "../../lib/validation";
-import { getUser, UserHuman, UserLab } from '../database/database';
-import JsonSchema, { schemas } from "../jsonSchemas/JsonSchema";
-import utils from '../utils'; 
+import { getModelForRole, getUserById } from '../database/database';
 import { IUserCommon } from '../database/schemas/IUserCommon';
+import JsonSchema, { schemaForRole } from "../jsonSchemas/JsonSchema";
+import utils from '../utils';
 
 class Profile {
     public async get(req: express.Request, res: express.Response, next: express.NextFunction) {
         let token = utils.getUnverifiedDecodedJWT(req);
         
-        let user = await getUser({ _id: token.sub });
+        let user = await getUserById(token.sub);
         if (!user) {
             return utils.badRequest(res);
         }
@@ -33,29 +32,24 @@ class Profile {
     async post(req: express.Request, res: express.Response, next: express.NextFunction) {
         let body: IUserCommon = req.body;
         let token = utils.getUnverifiedDecodedJWT(req);
-        let schema: any;
         
         delete body.password
         delete body._id
         delete body.__v
-        let model: Model<any>;
-        if (token.role === "human") {
-            // deep copy schema
-            schema = JSON.parse(JSON.stringify(schemas.registration_human));
-            model = UserHuman
-        }
-        else {
-            // deep copy schema
-            schema = JSON.parse(JSON.stringify(schemas.registration_lab));
-            model = UserLab
-        }
+        
+        let model = getModelForRole(token.role)
+        let schema = schemaForRole(token.role)
+        if (!model || !schema) return utils.badRequest(res);
+
         delete schema.required
 
-        if (!JsonSchema.validate(body, schema) || !v.validProfileFields(body, token.role).valid) {
-            return utils.badRequest(res);
+        let valResult = v.validProfileFields(body, token.role)
+        if (!JsonSchema.validate(body, schema) || !valResult.valid) {
+            if (!valResult.valid) return utils.handleError(res, valResult.err);
+            return utils.badRequest(res)
         }
 
-        let user = await getUser({ _id: token.sub })
+        let user = await getUserById(token.sub)
         if (!user) {
             return utils.badRequest(res)
         }
@@ -83,12 +77,9 @@ class Profile {
 
     async delete(req: express.Request, res: express.Response, next: express.NextFunction) {
         let token = utils.getUnverifiedDecodedJWT(req);
-        let model: mongoose.Model<any>;
-        if (token.role === "human") {
-            model = UserHuman;
-        }
-        else {
-            model = UserLab;
+        let model = getModelForRole(token.role)
+        if (!model) {
+            return utils.badRequest(res)
         }
 
         model.deleteOne({ _id: token.sub }).exec().then((doc) => {
